test(store): share fakeFirestore(), add a cross-implementation parity harness
Extracts fakeFirestore() out of store-firestore.test.ts into
store/fake-firestore.ts so it can be reused rather than redefined --
it was already the one piece of infrastructure the store split needs
and didn't have: InMemoryStore and FirestoreStore had almost no
overlapping test coverage (store.test.ts covers job transitions,
dispatch, publications, quota only against InMemoryStore;
store-firestore.test.ts covers saves, worlds, assessments, creation
limits only against the fake). Neither implementation's behaviour was
ever checked against the other's.

Adds store/store-parity.test.ts: a describeStoreContract() harness
that runs one shared assertion set against both InMemoryStore and
FirestoreStore(fakeFirestore().db). Proven on the oauth and telemetry
slices, per the settled Phase 2 plan -- the smallest and cleanest, and
a real test of whether the approach holds before betting Wave 4 on it.

Writing the harness surfaced three real gaps in the fake, each fixed
with production Firestore semantics rather than papered over:
  - .create() was unimplemented; several OAuth/access-token writes use
    it specifically because it rejects an existing document, unlike
    .set(). Added on doc refs, batches and transactions, with the
    ALREADY_EXISTS behaviour that distinction exists for.
  - .limit() only worked chained after .where() -- listTelemetryEvents
    calls it directly on the bare collection when no slug filter
    applies. Added to the collection object too.
  - .where() only supported '==' -- listVisitEvents' excludeType
    filter uses '!='. Added.
  - tx.update() was unimplemented; revokeOAuthGrant and
    rotateOAuthRefreshToken both use it inside a transaction. Added on
    both the transaction and the batch (already used elsewhere via
    markNotificationsRead).

store-firestore.test.ts: 710 -> 530 lines (the extracted fake moved
out); no baseline change needed since shrinking is always allowed.

Phase 2 wave 3 of the store-split plan (docs/phase2-store-split-
execution-plan.md in the private ops repo) -- closes waves 0-3, the
whole zero-behavioural-risk PR per the settled PR-shape decision.
